refactor: realoca o código dos arquivos
