+++ src/repositories/implementations/patients/index.ts
@@ -1,26 +1,65 @@
+import { PrismaClient } from "@prisma/client";
 import {
   IPatient,
   IPatientByIdReq,
   IPatientCreateReq,
 } from "../../../interfaces/patient.interface";
+import { IPatientRepository } from "../IPatientRepository";
 
-interface IPatientRepository {
-  createPatient({
+class PatientRepository implements IPatientRepository {
+  prisma: PrismaClient;
+
+  constructor() {
+    this.prisma = new PrismaClient();
+  }
+
+  public async createPatient({
     name,
     medical_records_number,
     contact,
-  }: IPatientCreateReq): Promise<IPatient>;
+  }: IPatientCreateReq): Promise<IPatient> {
+    const newPatient = await this.prisma.patient.create({
+      data: {
+        name,
+        medical_records_number,
+        contact,
+      },
+    });
 
-  getAllPatients(): Promise<IPatient[]>;
+    return newPatient;
+  }
 
-  getPatientById({ patient_id }: IPatientByIdReq): Promise<IPatient | null>;
+  public async getAllPatients(): Promise<IPatient[]> {
+    const allPatients = await this.prisma.patient.findMany();
+    return allPatients;
+  }
 
-  patchPatientById(
+  public async getPatientById({
+    patient_id,
+  }: IPatientByIdReq): Promise<IPatient | null> {
+    const patient = await this.prisma.patient.findUnique({
+      where: { id: patient_id },
+    });
+
+    return patient;
+  }
+
+  public async patchPatientById(
     { patient_id }: IPatientByIdReq,
     data: IPatientCreateReq
-  ): Promise<void>;
+  ): Promise<void> {
+    // await this.prisma.patient.update({
+    //   where: { id: patient_id },
+    //   data: { ...data },
+    // });
+    console.log(patient_id, data);
+  }
 
-  deletePatientById({ patient_id }: IPatientByIdReq): Promise<void>;
+  public async deletePatientById({
+    patient_id,
+  }: IPatientByIdReq): Promise<void> {
+    await this.prisma.patient.delete({ where: { id: patient_id } });
+  }
 }
 
-export { IPatientRepository };
+export { PatientRepository };
